pyproject.toml
[project]
name = "vt-trades-polymarket"
version = "0.1.0"
description = "Polymarket arbitrage trading bot — OpenClaw skill for VT Trades"
requires-python = ">=3.11"
dependencies = [
    "py-clob-client>=0.34.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
    "requests>=2.31.0",
    "web3>=6.15.0",
    "click>=8.1.0",
    "rich>=13.7.0",
    "httpx>=0.27.0",
]

[project.scripts]
polyclaw = "scripts.polyclaw:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
